% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{guidance_qmd_website}
\alias{guidance_qmd_website}
\title{Guidance lines for an existing standalone \verb{qmd/} website}
\usage{
guidance_qmd_website()
}
\description{
Guidance lines for an existing standalone \verb{qmd/} website
}
\keyword{internal}
