% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{guidance_quarto}
\alias{guidance_quarto}
\title{Guidance lines for a root \verb{_quarto.yml}}
\usage{
guidance_quarto(file, wired)
}
\description{
Guidance lines for a root \verb{_quarto.yml}
}
\keyword{internal}
