% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{line_references_qmd}
\alias{line_references_qmd}
\title{Does a single (comment-stripped) line reference \code{qmd} as a sequence element?}
\usage{
line_references_qmd(line)
}
\description{
Does a single (comment-stripped) line reference \code{qmd} as a sequence element?
}
\keyword{internal}
