% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmdme-package.R
\docType{package}
\name{qmdme-package}
\alias{qmdme}
\alias{qmdme-package}
\title{qmdme: Generate Quarto Code Companions for Source Files}
\description{
Produces and updates 'eval=FALSE' Quarto companion files for source code in a research repository. Each source file gets a 'qmd/' file with a user-editable prose region above an auto-regenerated code chunk, suitable for inclusion in a Quarto site.
}
\author{
\strong{Maintainer}: Jesse Brandt \email{jesse.a.brandt@gmail.com}

Authors:
\itemize{
  \item Jesse Brandt \email{jesse.a.brandt@gmail.com}
}

}
\keyword{internal}
