% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{qmdme_default_extensions}
\alias{qmdme_default_extensions}
\title{Default mapping of source-file extensions to Quarto chunk-language tags}
\usage{
qmdme_default_extensions()
}
\value{
Named character vector. Names are file extensions (without the dot);
values are Quarto chunk-language tags.
}
\description{
Default mapping of source-file extensions to Quarto chunk-language tags
}
\keyword{internal}
