% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{site_result}
\alias{site_result}
\title{Assemble a \code{detect_site()} result list}
\usage{
site_result(type, file, wired, guidance)
}
\description{
Assemble a \code{detect_site()} result list
}
\keyword{internal}
